Move the scanChanges function into it's own file
